python/pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.10.19"
dependencies = ["fastmcp==2.13.3", "oxygent==1.0.8"]
